Add files via upload
I am not able to add the project which I did in VS code direcctly to github , as some error is coming. Hence I have manually added the files. Unfortunately I am not able to show the fact that I have a sub folder named 'Scraper' in the main folder 'Crypto_Scraper'.